removed refactored_model
